Move ai_assistant app into poradnia/ to match app layout convention
Every other Django app lives under poradnia/; ai_assistant was the lone
exception. The app label stays "ai_assistant" (AppConfig.name changes
but no explicit label is set), so no new migrations are required.
