+++ config/urls.py
@@ -29,7 +29,9 @@ urlpatterns = [
     path("porady/", include("poradnia.advicer.urls", namespace="advicer")),
     path("uwagi/", include("poradnia.tasty_feedback.urls", namespace="tasty_feedback")),
     path("sprawy_sadowe/", include("poradnia.judgements.urls", namespace="judgements")),
-    path("ai-assistant/", include("ai_assistant.urls", namespace="ai_assistant")),
+    path(
+        "ai-assistant/", include("poradnia.ai_assistant.urls", namespace="ai_assistant")
+    ),
     path("teryt/", include("poradnia.teryt.urls", namespace="teryt")),
     path("robots.txt", TemplateView.as_view(template_name="robots.txt")),
     path(
