+++ poradnia/ai_assistant/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class AiAssistantConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "ai_assistant"
+    name = "poradnia.ai_assistant"
