+++ poradnia/ai_assistant/models.py
